+++ Content.Client/Species/Harpy/Components/HarpyVisualsComponent.cs
@@ -1,4 +1,4 @@
-namespace Content.Client.DeltaV.Harpy;
+namespace Content.Client.Species.Harpy.Components;
 
 [RegisterComponent]
 public sealed partial class HarpyVisualsComponent : Component
