+++ Content.Client/Species/Vulpkanin/Systems/DogVisionOverlay.cs
@@ -2,9 +2,9 @@ using Robust.Client.Graphics;
 using Robust.Client.Player;
 using Robust.Shared.Enums;
 using Robust.Shared.Prototypes;
-using Content.Shared.Abilities;
+using Content.Shared.Species.Vulpkanin.Components;
 
-namespace Content.Client.Nyanotrasen.Overlays;
+namespace Content.Client.Species.Vulpkanin.Systems;
 
 public sealed partial class DogVisionOverlay : Overlay
 {
