+++ Content.Server/Species/Oni/Components/HeldbyOniComponent.cs
@@ -1,4 +1,4 @@
-namespace Content.Server.Abilities.Oni
+namespace Content.Server.Species.Oni.Components
 {
     [RegisterComponent]
     public sealed partial class HeldByOniComponent : Component
