+++ Content.Shared/Item/ItemComponent.cs
@@ -1,5 +1,5 @@
 using Content.Shared.Hands.Components;
-using Content.Shared.Nyanotrasen.Item.PseudoItem;
+using Content.Shared.Species.Felinids.Systems;
 using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Serialization;
