+++ Content.Shared/Species/Diona/Components/NymphComponent.cs
@@ -1,10 +1,10 @@
 using Robust.Shared.Prototypes;
 using Robust.Shared.GameStates;
 
-namespace Content.Shared.Species.Components;
+namespace Content.Shared.Species.Diona.Components;
 /// <summary>
 /// This will replace one entity with another entity when it is removed from a body part.
-/// Obviously hyper-specific. If you somehow find another use for this, good on you. 
+/// Obviously hyper-specific. If you somehow find another use for this, good on you.
 /// </summary>
 
 [RegisterComponent, NetworkedComponent]
