+++ Content.Shared/Species/Felinids/Systems/HairballActionEvent.cs
@@ -1,3 +1,5 @@
-namespace Content.Shared.Actions.Events;
+using Content.Shared.Actions;
+
+namespace Content.Shared.Species.Felinids.Systems;
 
 public sealed partial class HairballActionEvent : InstantActionEvent {}
