+++ Content.Shared/Species/Harpy/Components/SharedHarpyVisualsComponent.cs
@@ -1,6 +1,6 @@
 using Robust.Shared.Serialization;
 
-namespace Content.Shared.DeltaV.Harpy;
+namespace Content.Shared.Species.Harpy.Components;
 
 [Serializable, NetSerializable]
 public enum HardsuitWings : byte
