+++ Content.Shared/Species/Spider/Components/IgnoreSpiderWebComponent.cs
@@ -1,4 +1,4 @@
-namespace Content.Shared.Spider;
+namespace Content.Shared.Species.Spider.Components;
 
 [RegisterComponent]
 public sealed partial class IgnoreSpiderWebComponent : Component
