+++ Content.Shared/Species/Spider/Components/SpiderWebObjectComponent.cs
@@ -1,6 +1,7 @@
+using Content.Shared.Species.Spider.Systems;
 using Robust.Shared.GameStates;
 
-namespace Content.Shared.Spider;
+namespace Content.Shared.Species.Spider.Components;
 
 [RegisterComponent, NetworkedComponent]
 [Access(typeof(SharedSpiderSystem))]
